data: support ClickHouse back (#874)
(cherry picked from commit 76a3ca5c56cc28a7d5dfe602c99949c24a9fa317)

+++ storage/data/sql.go
@@ -26,6 +26,7 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/juju/errors"
 	_ "github.com/lib/pq"
+	_ "github.com/mailru/go-clickhouse/v2"
 	"github.com/samber/lo"
 	"github.com/zhenghaoz/gorse/base/jsonutil"
 	"github.com/zhenghaoz/gorse/base/log"
@@ -42,6 +43,7 @@ type SQLDriver int
 const (
 	MySQL SQLDriver = iota
 	Postgres
+	ClickHouse
 	SQLite
 )
 
@@ -87,12 +89,53 @@ func NewSQLUser(user User) (sqlUser SQLUser) {
 	return
 }
 
+type ClickHouseItem struct {
+	SQLItem `gorm:"embedded"`
+	Version time.Time `gorm:"column:version"`
+}
+
+func NewClickHouseItem(item Item) (clickHouseItem ClickHouseItem) {
+	clickHouseItem.SQLItem = NewSQLItem(item)
+	clickHouseItem.Timestamp = item.Timestamp.In(time.UTC)
+	clickHouseItem.Version = time.Now().In(time.UTC)
+	return
+}
+
+type ClickhouseUser struct {
+	SQLUser `gorm:"embedded"`
+	Version time.Time `gorm:"column:version"`
+}
+
+func NewClickhouseUser(user User) (clickhouseUser ClickhouseUser) {
+	clickhouseUser.SQLUser = NewSQLUser(user)
+	clickhouseUser.Version = time.Now().In(time.UTC)
+	return
+}
+
+type ClickHouseFeedback struct {
+	Feedback `gorm:"embedded"`
+	Version  time.Time `gorm:"column:version"`
+}
+
 // SQLDatabase use MySQL as data storage.
 type SQLDatabase struct {
 	storage.TablePrefix
 	gormDB *gorm.DB
 	client *sql.DB
 	driver SQLDriver
+}
+
+// Optimize is used by ClickHouse only.
+func (d *SQLDatabase) Optimize() error {
+	if d.driver == ClickHouse {
+		for _, tableName := range []string{d.UsersTable(), d.ItemsTable(), d.FeedbackTable()} {
+			_, err := d.client.Exec("OPTIMIZE TABLE " + tableName)
+			if err != nil {
+				return errors.Trace(err)
+			}
+		}
+	}
+	return nil
 }
 
 // Init tables and indices in MySQL.
@@ -185,6 +228,44 @@ func (d *SQLDatabase) Init() error {
 		if err != nil {
 			return errors.Trace(err)
 		}
+	case ClickHouse:
+		// create tables
+		type Items struct {
+			ItemId     string    `gorm:"column:item_id;type:String"`
+			IsHidden   int       `gorm:"column:is_hidden;type:Boolean;default:0"`
+			Categories string    `gorm:"column:categories;type:String;default:'[]'"`
+			Timestamp  time.Time `gorm:"column:time_stamp;type:Datetime"`
+			Labels     string    `gorm:"column:labels;type:String;default:'[]'"`
+			Comment    string    `gorm:"column:comment;type:String"`
+			Version    struct{}  `gorm:"column:version;type:DateTime"`
+		}
+		err := d.gormDB.Set("gorm:table_options", "ENGINE = ReplacingMergeTree(version) ORDER BY item_id").AutoMigrate(Items{})
+		if err != nil {
+			return errors.Trace(err)
+		}
+		type Users struct {
+			UserId    string   `gorm:"column:user_id;type:String"`
+			Labels    string   `gorm:"column:labels;type:String;default:'[]'"`
+			Subscribe string   `gorm:"column:subscribe;type:String;default:'[]'"`
+			Comment   string   `gorm:"column:comment;type:String"`
+			Version   struct{} `gorm:"column:version;type:DateTime"`
+		}
+		err = d.gormDB.Set("gorm:table_options", "ENGINE = ReplacingMergeTree(version) ORDER BY user_id").AutoMigrate(Users{})
+		if err != nil {
+			return errors.Trace(err)
+		}
+		type Feedback struct {
+			FeedbackType string    `gorm:"column:feedback_type;type:String"`
+			UserId       string    `gorm:"column:user_id;type:String;index:user_index,type:bloom_filter(0.01),granularity:1"`
+			ItemId       string    `gorm:"column:item_id;type:String;index:item_index,type:bloom_filter(0.01),granularity:1"`
+			Timestamp    time.Time `gorm:"column:time_stamp;type:DateTime"`
+			Comment      string    `gorm:"column:comment;type:String"`
+			Version      struct{}  `gorm:"column:version;type:DateTime"`
+		}
+		err = d.gormDB.Set("gorm:table_options", "ENGINE = ReplacingMergeTree(version) ORDER BY (feedback_type, user_id, item_id)").AutoMigrate(Feedback{})
+		if err != nil {
+			return errors.Trace(err)
+		}
 	}
 	return nil
 }
@@ -200,10 +281,19 @@ func (d *SQLDatabase) Close() error {
 
 func (d *SQLDatabase) Purge() error {
 	tables := []string{d.ItemsTable(), d.FeedbackTable(), d.UsersTable()}
-	for _, tableName := range tables {
-		err := d.gormDB.Exec(fmt.Sprintf("DELETE FROM %s", tableName)).Error
-		if err != nil {
-			return errors.Trace(err)
+	if d.driver == ClickHouse {
+		for _, tableName := range tables {
+			err := d.gormDB.Exec(fmt.Sprintf("alter table %s delete where 1=1", tableName)).Error
+			if err != nil {
+				return errors.Trace(err)
+			}
+		}
+	} else {
+		for _, tableName := range tables {
+			err := d.gormDB.Exec(fmt.Sprintf("DELETE FROM %s", tableName)).Error
+			if err != nil {
+				return errors.Trace(err)
+			}
 		}
 	}
 	return nil
@@ -231,6 +321,36 @@ func (d *SQLDatabase) BatchInsertItems(ctx context.Context, items []Item) error 
 		DoUpdates: clause.AssignmentColumns([]string{"is_hidden", "categories", "time_stamp", "labels", "comment"}),
 	}).Create(rows).Error
 	return errors.Trace(err)
+	if d.driver == ClickHouse {
+		rows := make([]ClickHouseItem, 0, len(items))
+		memo := mapset.NewSet[string]()
+		for _, item := range items {
+			if !memo.Contains(item.ItemId) {
+				memo.Add(item.ItemId)
+				rows = append(rows, NewClickHouseItem(item))
+			}
+		}
+		err := d.gormDB.Create(rows).Error
+		return errors.Trace(err)
+	} else {
+		rows := make([]SQLItem, 0, len(items))
+		memo := mapset.NewSet[string]()
+		for _, item := range items {
+			if !memo.Contains(item.ItemId) {
+				memo.Add(item.ItemId)
+				row := NewSQLItem(item)
+				if d.driver == SQLite {
+					row.Timestamp = row.Timestamp.In(time.UTC)
+				}
+				rows = append(rows, row)
+			}
+		}
+		err := d.gormDB.WithContext(ctx).Clauses(clause.OnConflict{
+			Columns:   []clause.Column{{Name: "item_id"}},
+			DoUpdates: clause.AssignmentColumns([]string{"is_hidden", "categories", "time_stamp", "labels", "comment"}),
+		}).Create(rows).Error
+		return errors.Trace(err)
+	}
 }
 
 func (d *SQLDatabase) BatchGetItems(ctx context.Context, namespace string, itemIds []string) ([]Item, error) {
@@ -319,7 +439,7 @@ func (d *SQLDatabase) ModifyItem(ctx context.Context, namespace string, itemId s
 	}
 	if patch.Timestamp != nil {
 		switch d.driver {
-		case SQLite:
+		case ClickHouse, SQLite:
 			attributes["time_stamp"] = patch.Timestamp.In(time.UTC)
 		default:
 			attributes["time_stamp"] = patch.Timestamp
@@ -452,19 +572,32 @@ func (d *SQLDatabase) BatchInsertUsers(ctx context.Context, users []User) error 
 	if len(users) == 0 {
 		return nil
 	}
-	rows := make([]SQLUser, 0, len(users))
-	memo := mapset.NewSet[string]()
-	for _, user := range users {
-		if !memo.Contains(user.UserId) {
-			memo.Add(user.UserId)
-			rows = append(rows, NewSQLUser(user))
+	if d.driver == ClickHouse {
+		rows := make([]ClickhouseUser, 0, len(users))
+		memo := mapset.NewSet[string]()
+		for _, user := range users {
+			if !memo.Contains(user.UserId) {
+				memo.Add(user.UserId)
+				rows = append(rows, NewClickhouseUser(user))
+			}
 		}
+		err := d.gormDB.Create(rows).Error
+		return errors.Trace(err)
+	} else {
+		rows := make([]SQLUser, 0, len(users))
+		memo := mapset.NewSet[string]()
+		for _, user := range users {
+			if !memo.Contains(user.UserId) {
+				memo.Add(user.UserId)
+				rows = append(rows, NewSQLUser(user))
+			}
+		}
+		err := d.gormDB.WithContext(ctx).Clauses(clause.OnConflict{
+			Columns:   []clause.Column{{Name: "user_id"}},
+			DoUpdates: clause.AssignmentColumns([]string{"labels", "subscribe", "comment"}),
+		}).Create(rows).Error
+		return errors.Trace(err)
 	}
-	err := d.gormDB.WithContext(ctx).Clauses(clause.OnConflict{
-		Columns:   []clause.Column{{Name: "user_id"}},
-		DoUpdates: clause.AssignmentColumns([]string{"labels", "subscribe", "comment"}),
-	}).Create(rows).Error
-	return errors.Trace(err)
 }
 
 // DeleteUser deletes a user from MySQL.
@@ -634,18 +767,33 @@ func (d *SQLDatabase) BatchInsertFeedback(ctx context.Context, feedback []Feedba
 	// insert users
 	if insertUser {
 		userList := users.ToSlice()
-		err := tx.Clauses(clause.OnConflict{
-			Columns:   []clause.Column{{Name: "user_id"}},
-			DoNothing: true,
-		}).Create(lo.Map(userList, func(userId string, _ int) SQLUser {
-			return SQLUser{
-				UserId:    userId,
-				Labels:    "null",
-				Subscribe: "null",
+		if d.driver == ClickHouse {
+			err := tx.Create(lo.Map(userList, func(userId string, _ int) ClickhouseUser {
+				return ClickhouseUser{
+					SQLUser: SQLUser{
+						UserId:    userId,
+						Labels:    "[]",
+						Subscribe: "[]",
+					},
+				}
+			})).Error
+			if err != nil {
+				return errors.Trace(err)
 			}
-		})).Error
-		if err != nil {
-			return errors.Trace(err)
+		} else {
+			err := tx.Clauses(clause.OnConflict{
+				Columns:   []clause.Column{{Name: "user_id"}},
+				DoNothing: true,
+			}).Create(lo.Map(userList, func(userId string, _ int) SQLUser {
+				return SQLUser{
+					UserId:    userId,
+					Labels:    "null",
+					Subscribe: "null",
+				}
+			})).Error
+			if err != nil {
+				return errors.Trace(err)
+			}
 		}
 	} else {
 		for _, user := range users.ToSlice() {
@@ -679,6 +827,33 @@ func (d *SQLDatabase) BatchInsertFeedback(ctx context.Context, feedback []Feedba
 		})).Error
 		if err != nil {
 			return errors.Trace(err)
+		if d.driver == ClickHouse {
+			err := tx.Create(lo.Map(itemList, func(itemId string, _ int) ClickHouseItem {
+				return ClickHouseItem{
+					SQLItem: SQLItem{
+						ItemId:     itemId,
+						Labels:     "[]",
+						Categories: "[]",
+					},
+				}
+			})).Error
+			if err != nil {
+				return errors.Trace(err)
+			}
+		} else {
+			err := tx.Clauses(clause.OnConflict{
+				Columns:   []clause.Column{{Name: "item_id"}},
+				DoNothing: true,
+			}).Create(lo.Map(itemList, func(itemId string, _ int) SQLItem {
+				return SQLItem{
+					ItemId:     itemId,
+					Labels:     "null",
+					Categories: "null",
+				}
+			})).Error
+			if err != nil {
+				return errors.Trace(err)
+			}
 		}
 	} else {
 		for _, item := range items.ToSlice() {
@@ -723,6 +898,50 @@ func (d *SQLDatabase) BatchInsertFeedback(ctx context.Context, feedback []Feedba
 		DoUpdates: lo.If(overwrite, clause.AssignmentColumns([]string{"time_stamp", "comment"})).Else(nil),
 	}).Create(rows).Error
 	return errors.Trace(err)
+	if d.driver == ClickHouse {
+		rows := make([]ClickHouseFeedback, 0, len(feedback))
+		memo := make(map[lo.Tuple3[string, string, string]]struct{})
+		for _, f := range feedback {
+			if users.Contains(f.UserId) && items.Contains(f.ItemId) {
+				if _, exist := memo[lo.Tuple3[string, string, string]{f.FeedbackType, f.UserId, f.ItemId}]; !exist {
+					memo[lo.Tuple3[string, string, string]{f.FeedbackType, f.UserId, f.ItemId}] = struct{}{}
+					f.Timestamp = f.Timestamp.In(time.UTC)
+					rows = append(rows, ClickHouseFeedback{
+						Feedback: f,
+						Version:  lo.If(overwrite, time.Now().In(time.UTC)).Else(time.Time{}),
+					})
+				}
+			}
+		}
+		if len(rows) == 0 {
+			return nil
+		}
+		err := tx.Create(rows).Error
+		return errors.Trace(err)
+	} else {
+		rows := make([]Feedback, 0, len(feedback))
+		memo := make(map[lo.Tuple3[string, string, string]]struct{})
+		for _, f := range feedback {
+			if users.Contains(f.UserId) && items.Contains(f.ItemId) {
+				if _, exist := memo[lo.Tuple3[string, string, string]{f.FeedbackType, f.UserId, f.ItemId}]; !exist {
+					memo[lo.Tuple3[string, string, string]{f.FeedbackType, f.UserId, f.ItemId}] = struct{}{}
+					if d.driver == SQLite {
+						f.Timestamp = f.Timestamp.In(time.UTC)
+					}
+					rows = append(rows, f)
+				}
+			}
+		}
+		if len(rows) == 0 {
+			return nil
+		}
+		err := tx.Clauses(clause.OnConflict{
+			Columns:   []clause.Column{{Name: "feedback_type"}, {Name: "user_id"}, {Name: "item_id"}},
+			DoNothing: !overwrite,
+			DoUpdates: lo.If(overwrite, clause.AssignmentColumns([]string{"time_stamp", "comment"})).Else(nil),
+		}).Create(rows).Error
+		return errors.Trace(err)
+	}
 }
 
 // GetFeedback returns feedback from MySQL.
@@ -865,7 +1084,7 @@ func (d *SQLDatabase) DeleteUserItemFeedback(ctx context.Context, namespace, use
 	if tx.Error != nil {
 		return 0, errors.Trace(tx.Error)
 	}
-	if tx.Error != nil {
+	if tx.Error != nil && d.driver != ClickHouse {
 		return 0, errors.Trace(tx.Error)
 	}
 	return int(tx.RowsAffected), nil
@@ -873,7 +1092,7 @@ func (d *SQLDatabase) DeleteUserItemFeedback(ctx context.Context, namespace, use
 
 func (d *SQLDatabase) convertTimeZone(timestamp *time.Time) time.Time {
 	switch d.driver {
-	case SQLite:
+	case ClickHouse, SQLite:
 		return timestamp.In(time.UTC)
 	default:
 		return *timestamp
